Adds executables and integration tests for AES-CTR and json files
git-svn-id: https://subversion.nist.gov/JIRA_repo1/STVM_Research/ACVP/gen-val/trunk@391281 58cf25f4-cab2-a041-9f61-d3785993b9bc

+++ common/test/NIST.CVP.Tests.Core/GenValTestsBase.cs
@@ -41,6 +41,9 @@ namespace NIST.CVP.Tests.Core
 
         public abstract void SetUp();
 
+        // Set this during a test if you want to save the json from the session
+        public bool SaveJson = false;
+
         [OneTimeSetUp]
         public void OneTimeSetUp()
         {
@@ -51,7 +54,10 @@ namespace NIST.CVP.Tests.Core
         [OneTimeTearDown]
         public void OneTimeTearDown()
         {
-            Directory.Delete(TestPath, true);
+            if (!SaveJson)
+            {
+                Directory.Delete(TestPath, true);
+            }
         }
 
         [Test]
